refactoring : navigation
